refactor(card): 카드 디자인 선택 페이지 구조 수정(components&pages 분리)
