+++ src/components/card/CardDesign.css
@@ -1,0 +1,268 @@
+/* CardDesign.css */
+
+/* 전체 컨테이너 */
+.card-design-container {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  background-color: #ffffff;
+  font-family: "Apple SD Gothic Neo", "Noto Sans KR", sans-serif;
+  position: relative;
+}
+
+/* 컨텐츠 영역 */
+.card-design-content {
+  flex: 1;
+  overflow-y: auto;
+  padding-bottom: 100px; /* 메뉴 공간 확보 */
+}
+
+/* 타이틀 영역 */
+.card-design-title {
+  padding: 24px 20px;
+  text-align: center;
+}
+
+.card-design-title h2 {
+  font-size: 22px;
+  font-weight: bold;
+  margin: 0;
+  color: #333;
+}
+
+/* 카드 프리뷰 영역 */
+.card-preview {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 0 20px;
+  margin-bottom: 40px;
+}
+
+/* 카드 이미지 스타일 */
+.card-image {
+  width: 190px;
+  height: 300px;
+  border-radius: 16px;
+  position: relative;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  display: flex;
+  flex-direction: column;
+  padding: 20px;
+  margin-bottom: 20px;
+  transition: transform 0.3s;
+  transform-style: preserve-3d;
+  perspective: 1000px;
+}
+
+/* 카드 뒤집기 애니메이션 */
+.card-image.flipping {
+  transform: rotateY(180deg);
+}
+
+/* 기본 카드 앞면 */
+.card-image.frontBasic {
+  background: linear-gradient(to bottom, #ffcccc, #ffffff);
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  position: relative;
+}
+
+/* 기본 카드 뒷면 */
+.card-image.backBasic {
+  background: linear-gradient(to bottom, #f0f0f0, #dddddd);
+  transform: rotateY(180deg);
+}
+
+/* 커스텀 카드 */
+.card-image.frontCustom {
+  background: linear-gradient(to bottom, #c4e0ff, #ffffff);
+  position: relative;
+}
+
+/* 카드 뒷면 내용 */
+.card-back-content {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  transform: rotateY(180deg);
+}
+
+.signature-area {
+  width: 90%;
+  height: 40px;
+  background-color: #fff;
+  border-radius: 4px;
+  margin: 20px auto;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #999;
+  font-size: 14px;
+}
+
+.card-back-info {
+  margin-top: auto;
+  text-align: center;
+  font-size: 12px;
+  color: #555;
+}
+
+.card-back-info p {
+  margin: 4px 0;
+}
+
+.card-chip {
+  width: 40px;
+  height: 30px;
+  background-color: #cccccc;
+  border-radius: 5px;
+  position: absolute;
+  top: 40%;
+  right: 20px;
+}
+
+.card-number {
+  position: absolute;
+  top: 55%;
+  width: 100%;
+  text-align: center;
+  font-size: 18px;
+  font-weight: 500;
+  letter-spacing: 2px;
+  color: #333;
+  left: 0;
+}
+
+.card-info {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  background-color: #333333;
+  color: white;
+  padding: 15px 20px;
+  border-bottom-left-radius: 16px;
+  border-bottom-right-radius: 16px;
+}
+
+.card-holder {
+  font-size: 16px;
+  font-weight: 500;
+}
+
+.card-expiry {
+  font-size: 14px;
+  margin-top: 4px;
+}
+
+/* 화살표 버튼 스타일 */
+.card-arrow {
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  font-size: 28px;
+  font-weight: bold;
+  color: #333;
+  background-color: #fff;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  cursor: pointer;
+  transition: transform 0.2s;
+  z-index: 10;
+}
+
+.card-next-arrow {
+  right: -20px;
+}
+
+.card-prev-arrow {
+  left: -20px;
+}
+
+.card-arrow:hover {
+  transform: translateY(-50%) scale(1.05);
+}
+
+/* 디자인 선택 영역 */
+.design-selection {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 20px;
+  text-align: center;
+  position: relative;
+}
+
+.design-label {
+  font-size: 16px;
+  font-weight: 500;
+  margin-bottom: 15px;
+  color: #333;
+  margin-top: 5px;
+}
+
+.design-button {
+  background-color: #fff;
+  border: 1px solid #ddd;
+  border-radius: 20px;
+  padding: 10px 30px;
+  font-size: 14px;
+  cursor: pointer;
+  color: #666;
+  transition: all 0.2s;
+}
+
+.design-button:hover {
+  background-color: #f8f8f8;
+}
+
+.custom-design-button {
+  background-color: #8470ff;
+  color: #ffffff;
+  border: none;
+}
+
+.custom-design-button:hover {
+  background-color: #7360ee;
+}
+
+/* 다음 버튼 */
+.next-button {
+  background-color: #6c5ce7;
+  color: white;
+  border: none;
+  border-radius: 25px;
+  padding: 12px 50px;
+  font-size: 16px;
+  font-weight: bold;
+  cursor: pointer;
+  transition: background-color 0.3s;
+  width: 80%;
+  margin: 0 auto;
+  margin-top: auto;
+  margin-bottom: 150px;
+}
+
+.next-button:hover {
+  background-color: #5b4bc4;
+}
+
+/* 반응형 스타일 */
+@media (max-width: 360px) {
+  .card-image {
+    width: 280px;
+    height: 180px;
+  }
+
+  .card-design-title h2 {
+    font-size: 20px;
+  }
+}
